+++ src/models/UserLogin.php
@@ -104,7 +104,7 @@ class UserLogin extends ActiveRecord
     /**
      * @inheritdoc
      */
-    public static function tableName()
+    public static function tableName(): string
     {
         return '{{%user_login}}';
     }