+++ src/web/DbSession.php
@@ -2,10 +2,6 @@
 
 namespace davidhirtz\yii2\skeleton\web;
 
-/**
- * Class DbSession
- * @package davidhirtz\yii2\skeleton\web
- */
 class DbSession extends \yii\web\DbSession
 {
     use SessionTrait;